requirements.txt
streamlit
torch
torchvision
pillow
timm
pandas
supabase
grad-cam
opencv-python-headless